+++ x/posts/internal/types/post_test.go
@@ -224,12 +224,12 @@ func TestPost_Validate(t *testing.T) {
 		},
 	}
 	answer := types.PollAnswer{
-		ID:   uint64(1),
+		ID:   uint(1),
 		Text: "Yes",
 	}
 
 	answer2 := types.PollAnswer{
-		ID:   uint64(2),
+		ID:   uint(2),
 		Text: "No",
 	}
 	pollData := types.NewPollData("poll?", time.Now().UTC().Add(time.Hour), types.PollAnswers{answer, answer2}, true, false, true)
@@ -393,8 +393,8 @@ func TestPost_Equals(t *testing.T) {
 		},
 	}
 
-	answer := types.PollAnswer{ID: uint64(1), Text: "Yes"}
-	answer2 := types.PollAnswer{ID: uint64(2), Text: "No"}
+	answer := types.PollAnswer{ID: uint(1), Text: "Yes"}
+	answer2 := types.PollAnswer{ID: uint(2), Text: "No"}
 	pollData := types.NewPollData("poll?", testPostEndPollDate, types.PollAnswers{answer, answer2}, true, false, true)
 
 	tests := []struct {
@@ -831,12 +831,12 @@ func TestPosts_String(t *testing.T) {
 		},
 	}
 	answer := types.PollAnswer{
-		ID:   uint64(1),
+		ID:   uint(1),
 		Text: "Yes",
 	}
 
 	answer2 := types.PollAnswer{
-		ID:   uint64(2),
+		ID:   uint(2),
 		Text: "No",
 	}
 	pollData := types.NewPollData("poll?", time.Now().UTC().Add(time.Hour), types.PollAnswers{answer, answer2}, true, false, true)
